+++ tests/sentry/api/endpoints/test_organization_project_slugs.py
